infra-modules-1/pyproject.toml
[tool.poetry]
name = "infra-modules"
version = "15.4.0"
description = "Infrastructure for the data-platform"
authors = ["Roel Smits <roel.smits@nn.nl> & Jeroen Sies <jeroen.sies@nn.nl>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0.1"
typer = "^0.12.3"
packaging = "^22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
